+++ common/core/src/main/java/is/codion/common/value/ValueList.java
@@ -36,7 +36,7 @@ public interface ValueList<T> extends Values<T, List<T>> {
 	ObservableValueList<T> observable();
 
 	/**
-	 * Creates a new empty {@link ValueList}, using {@link Notify#WHEN_CHANGED}.
+	 * Creates a new empty {@link ValueList}, using {@link Notify#CHANGED}.
 	 * @param <T> the value type
 	 * @return a new {@link ValueList}
 	 */
@@ -45,7 +45,7 @@ public interface ValueList<T> extends Values<T, List<T>> {
 	}
 
 	/**
-	 * Creates a new {@link ValueList}, using {@link Notify#WHEN_CHANGED}.
+	 * Creates a new {@link ValueList}, using {@link Notify#CHANGED}.
 	 * @param values the initial values, may not be null
 	 * @param <T> the value type
 	 * @return a new {@link ValueList}