+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 =================
 
 ## 0.18.35
+### is.codion.common.core
+- Value.Notify.WHEN_CHANGED and WHEN_SET renamed CHANGED and SET respectively.
 ### is.codion.common.db
 - Database.URL_SCOPED_INSTANCE configuration value added.
 ### is.codion.swing.common.model